+++ desktop/applis/apps/bean/bean_core/src/main/java/code/bean/nat/BeanNatCommonLgNames.java
@@ -237,10 +237,10 @@ public abstract class BeanNatCommonLgNames implements BeanNatCommonLgNamesInt, A
         }
         return arr_;
     }
-    public static NatArrayStruct getStrInteger(AbsMap<String, Integer> _map) {
+    public static NatArrayStruct getStrInteger(AbsMap<String, Long> _map) {
         NatArrayStruct arr_ = new NatArrayStruct(_map.size());
         int i_ = 0;
-        for (EntryCust<String,Integer> e: _map.entryList()){
+        for (EntryCust<String,Long> e: _map.entryList()){
             PairStruct p_ = new PairStruct(new NaStSt(StringUtil.nullToEmpty(e.getKey())),new NaNbSt(e.getValue()));
             arr_.set(i_,p_);
             i_++;
